one_person_company/services/ecommerce_service.py
"""
电商数据服务 - 抖音/1688/小红书数据获取
"""

import os
import json
import logging
import requests
from typing import Dict, List, Optional
from datetime import datetime
from bs4 import BeautifulSoup
import re

logger = logging.getLogger(__name__)


class EcommerceService:
    """
    电商数据服务
    支持：抖音热销、1688货源、小红书趋势
    """
    
    def __init__(self, config: Dict = None):
        self.config = config or {}
        self.headers = {
            "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36"
        }
        
    # ==================== 抖音数据 ====================
    
    def get_douyin_hot_products(self, category: str = None) -> Dict:
        """
        获取抖音热销产品
        通过公开数据和搜索分析
        """
        # 使用搜索引擎获取抖音热销信息
        from .search_service import SearchService
        search = SearchService()
        
        query = f"抖音 热销 爆款 {category or ''} 2024"
        result = search.search_web(query, count=10)
        
        hot_products = {
            "timestamp": datetime.now().isoformat(),
            "category": category,
            "source": "search_analysis",
            "products": []
        }
        
        if result.get("status") == "success":
            for item in result.get("results", []):
                # 从搜索结果提取产品信息
                hot_products["products"].append({
                    "title": item.get("title", ""),
                    "source_url": item.get("url", ""),
                    "description": item.get("description", ""),
                    "source": item.get("source", "")
                })
        
        return hot_products
    
    def analyze_douyin_product(self, product_name: str) -> Dict:
        """
        分析抖音产品数据
        """
        from .search_service import SearchService
        search = SearchService()
        
        # 多维度搜索
        queries = [
            f"{product_name} 抖音 销量",
            f"{product_name} 抖音 达人带货",
            f"{product_name} 抖音 价格"
        ]
        
        analysis = {
            "product": product_name,
            "timestamp": datetime.now().isoformat(),
            "douyin_data": {}
        }
        
        for query in queries:
            result = search.search_web(query, count=5)
            if result.get("status") == "success":
                analysis["douyin_data"][query] = result.get("results", [])
        
        return analysis
    
    # ==================== 1688数据 ====================
    
    def search_1688_suppliers(self, product_name: str, min_rating: float = 4.5) -> Dict:
        """
        搜索1688供应商
        """
        from .search_service import SearchService
        search = SearchService()
        
        query = f"1688 {product_name} 源头工厂 一件代发"
        result = search.search_web(query, count=10)
        
        suppliers = {
            "product": product_name,
            "timestamp": datetime.now().isoformat(),
            "platform": "1688",
            "suppliers": []
        }
        
        if result.get("status") == "success":
            for item in result.get("results", []):
                if "1688" in item.get("url", "") or "alibaba" in item.get("url", ""):
                    suppliers["suppliers"].append({
                        "title": item.get("title", ""),
                        "url": item.get("url", ""),
                        "description": item.get("description", "")
                    })
        
        return suppliers
    
    def get_1688_price_range(self, product_name: str) -> Dict:
        """
        获取1688价格区间
        """
        from .search_service import SearchService
        search = SearchService()
        
        query = f"1688 {product_name} 批发价格 报价"
        result = search.search_web(query, count=10)
        
        prices = {
            "product": product_name,
            "timestamp": datetime.now().isoformat(),
            "price_info": []
        }
        
        if result.get("status") == "success":
            for item in result.get("results", []):
                prices["price_info"].append({
                    "source": item.get("title", ""),
                    "url": item.get("url", ""),
                    "info": item.get("description", "")
                })
        
        return prices
    
    # ==================== 小红书数据 ====================
    
    def get_xiaohongshu_trends(self, category: str = None) -> Dict:
        """
        获取小红书趋势
        """
        from .search_service import SearchService
        search = SearchService()
        
        query = f"小红书 {category or ''} 热门 种草 2024"
        result = search.search_web(query, count=10)
        
        trends = {
            "timestamp": datetime.now().isoformat(),
            "category": category,
            "platform": "xiaohongshu",
            "trends": []
        }
        
        if result.get("status") == "success":
            for item in result.get("results", []):
                trends["trends"].append({
                    "title": item.get("title", ""),
                    "url": item.get("url", ""),
                    "description": item.get("description", "")
                })
        
        return trends
    
    # ==================== 综合分析 ====================
    
    def analyze_product_opportunity(self, product_name: str) -> Dict:
        """
        综合分析产品机会
        结合多平台数据
        """
        analysis = {
            "product": product_name,
            "timestamp": datetime.now().isoformat(),
            "platforms": {}
        }
        
        # 抖音数据
        analysis["platforms"]["douyin"] = self.get_douyin_hot_products(product_name)
        
        # 1688供应商
        analysis["platforms"]["1688"] = self.search_1688_suppliers(product_name)
        
        # 小红书趋势
        analysis["platforms"]["xiaohongshu"] = self.get_xiaohongshu_trends(product_name)
        
        # 综合评估
        analysis["assessment"] = self._assess_opportunity(analysis)
        
        return analysis
    
    def _assess_opportunity(self, analysis: Dict) -> Dict:
        """评估产品机会"""
        # 简单评估逻辑
        douyin_results = len(analysis.get("platforms", {}).get("douyin", {}).get("products", []))
        supplier_results = len(analysis.get("platforms", {}).get("1688", {}).get("suppliers", []))
        trend_results = len(analysis.get("platforms", {}).get("xiaohongshu", {}).get("trends", []))
        
        score = 0
        reasons = []
        
        if douyin_results >= 5:
            score += 30
            reasons.append("抖音热度高")
        
        if supplier_results >= 3:
            score += 30
            reasons.append("供应商充足")
        
        if trend_results >= 5:
            score += 20
            reasons.append("小红书有热度")
        
        return {
            "score": score,
            "max_score": 100,
            "reasons": reasons,
            "recommendation": "推荐" if score >= 50 else "需进一步分析"
        }
    
    def compare_suppliers(self, product_name: str) -> Dict:
        """
        比较多个供应商
        """
        suppliers_data = self.search_1688_suppliers(product_name)
        prices_data = self.get_1688_price_range(product_name)
        
        return {
            "product": product_name,
            "timestamp": datetime.now().isoformat(),
            "suppliers": suppliers_data.get("suppliers", []),
            "price_info": prices_data.get("price_info", []),
            "recommendation": "请进一步联系供应商获取详细报价"
        }


# 测试
if __name__ == "__main__":
    service = EcommerceService()
    
    # 测试产品机会分析
    result = service.analyze_product_opportunity("筋膜枪")
    print(json.dumps(result, ensure_ascii=False, indent=2))
