one_person_company/requirements.txt
# Python依赖
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
schedule>=1.2.0

# AI/LLM
anthropic>=0.18.0
openai>=1.12.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 金融数据
akshare>=1.12.0
yfinance>=0.2.36

# Web框架（仪表盘）
fastapi>=0.109.0
uvicorn>=0.27.0

# 数据库
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# HTML解析
beautifulsoup4>=4.12.0
lxml>=5.1.0

# 日志和监控
loguru>=0.7.0
rich>=13.0.0

# 测试
pytest>=8.0.0
pytest-asyncio>=0.23.0
